requirements.txt
Django==3.2.10
djangorestframework==3.12.4
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1